# api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entity.Designation:
    properties:
      created_at:
        type: string
      id:
        description: Primary key
        type: integer
      name:
        maxLength: 100
        minLength: 3
        type: string
      status:
        type: boolean
      updated_at:
        type: string
    required:
    - name
    type: object
  entity.DesignationResponsePagination:
    properties:
      data:
        items:
          $ref: '#/definitions/entity.ResponseDesignation'
        type: array
      pagination:
        $ref: '#/definitions/entity.Pagination'
    type: object
...
  entity.Pagination:
    properties:
      current_page:
        type: integer
      first_page:
        type: integer
      last_page:
        type: integer
      next_page:
        type: integer
      prev_page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
  entity.ResponseDesignation:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: boolean
      updated_at:
        type: string
    type: object
  entity.ResponseRole:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: boolean
      updated_at:
        type: string
    type: object
  entity.ResponseUser:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: boolean
      updated_at:
        type: string
    type: object
  entity.Role:
    properties:
      created_at:
        type: string
      id:
        description: Primary key
        type: integer
      name:
        maxLength: 100
        minLength: 3
        type: string
      status:
        type: boolean
      updated_at:
        type: string
    required:
    - name
    type: object
  entity.RoleResponsePagination:
    properties:
      data:
        items:
          $ref: '#/definitions/entity.ResponseRole'
        type: array
      pagination:
        $ref: '#/definitions/entity.Pagination'
    type: object
...
  entity.UpdateDesignation:
    properties:
      name:
        maxLength: 100
        minLength: 3
        type: string
      status:
        type: boolean
      updated_at:
        type: string
    type: object
  entity.UpdateRole:
    properties:
      name:
        maxLength: 100
        minLength: 3
        type: string
      status:
        type: boolean
      updated_at:
        type: string
    type: object
  entity.UpdateUser:
    properties:
      name:
        maxLength: 100
        minLength: 3
        type: string
      status:
        type: boolean
      updated_at:
        type: string
    type: object
  entity.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        description: Primary key
        type: integer
      name:
        maxLength: 100
        minLength: 3
        type: string
      password:
        maxLength: 100
        minLength: 6
        type: string
      role:
        type: integer
      status:
        type: boolean
      updated_at:
        type: string
    required:
    - email
    - name
    - password
    type: object
  entity.UserResponsePagination:
    properties:
      data:
        items:
          $ref: '#/definitions/entity.ResponseUser'
        type: array
      pagination:
        $ref: '#/definitions/entity.Pagination'
    type: object
host: localhost:9009
info:
  contact:
    email: jubaer01.cse@gmail.com
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API documentation for the News API
  title: News API
  version: "1.0"
paths:
...
  /designations:
    get:
      consumes:
      - application/json
      description: Get details of all designations
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      - description: Sort by
        in: query
        name: sort
        type: string
      - description: Filter by status
        in: query
        name: status
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.DesignationResponsePagination'
      security:
      - ApiKeyAuth: []
      summary: Get all designations
      tags:
      - designations
    post:
      consumes:
      - application/json
      description: Create a new Designation
      parameters:
      - description: The Designation to be created
        in: body
        name: designation
        required: true
        schema:
          $ref: '#/definitions/entity.Designation'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new Designation
      tags:
      - designations
  /designations/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Designation
      parameters:
      - description: The ID of the Designation
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a Designation
      tags:
      - designations
    put:
      consumes:
      - application/json
      description: Update an existing Designation
      parameters:
      - description: The ID of the Designation
        in: path
        name: id
        required: true
        type: string
      - description: Updated Designation object
        in: body
        name: designation
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateDesignation'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update an existing Designation
      tags:
      - designations
  /designations/{id}/details:
    get:
      consumes:
      - application/json
      description: Get detailed information about a Designation by ID
      parameters:
      - description: The ID of the Designation
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ResponseDesignation'
      security:
      - ApiKeyAuth: []
      summary: Get detailed information about a Designation by ID
      tags:
      - designations
  /roles:
    get:
      consumes:
      - application/json
      description: Get details of all roles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.RoleResponsePagination'
      summary: Get all roles
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: Create a new Role
      parameters:
      - description: The Role to be created
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/entity.Role'
      produces:
      - application/json
      responses: {}
      summary: Create a new Role
      tags:
      - roles
  /roles/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Role
      parameters:
      - description: The ID of the Role
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Delete a Role
      tags:
      - roles
    put:
      consumes:
      - application/json
      description: Update an existing Role
      parameters:
      - description: The ID of the Role
        in: path
        name: id
        required: true
        type: string
      - description: Updated Role object
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateRole'
      produces:
      - application/json
      responses: {}
      summary: Update an existing Role
      tags:
      - roles
  /roles/{id}/details:
    get:
      consumes:
      - application/json
      description: Get detailed information about a Role by ID
      parameters:
      - description: The ID of the Role
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get detailed information about a Role by ID
      tags:
      - roles
  /users:
    get:
      consumes:
      - application/json
      description: Get details of all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UserResponsePagination'
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new User
      parameters:
      - description: The User to be created
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entity.User'
      produces:
      - application/json
      responses: {}
      summary: Create a new User
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a User
      parameters:
      - description: The ID of the User
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Delete a User
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update an existing User
      parameters:
      - description: The ID of the User
        in: path
        name: id
        required: true
        type: string
      - description: Updated User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateUser'
      produces:
      - application/json
      responses: {}
      summary: Update an existing User
      tags:
      - users
  /users/{id}/details:
    get:
      consumes:
      - application/json
      description: Get detailed information about a User by ID
      parameters:
      - description: The ID of the User
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get detailed information about a User by ID
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"